feat(daengle): add guest api instance
