+++ apps/daengle/src/apis/config/index.ts
@@ -1,21 +1,16 @@
-import axios from 'axios';
-import { createHttpClient } from '@daengle/services/apis';
+import axios, { AxiosResponse } from 'axios';
+import { createHttpClient, HttpClient } from '@daengle/services/apis';
 
 export const { api } = createHttpClient({
   baseURL: process.env.NEXT_PUBLIC_API_URL ?? '',
   role: 'user',
 });
 
-export const guestApi = axios.create({
+export const guestApi: HttpClient = axios.create({
   baseURL: process.env.NEXT_PUBLIC_API_URL ?? '',
 });
 
-guestApi.interceptors.request.use(async (config) => {
-  const accessToken = localStorage.getItem('accessToken');
-
-  if (accessToken) {
-    config.headers['Authorization'] = `Bearer ${accessToken}`;
-  }
-
-  return config;
-});
+guestApi.interceptors.response.use(
+  (response: AxiosResponse) => response.data?.response,
+  (error) => Promise.reject(error)
+);
