+++ apps/daengle/src/apis/index.ts
@@ -2,5 +2,6 @@ export * from './auth';
 export * from './chat';
 export * from './config';
 export * from './estimate';
+export * from './home';
 export * from './payment';
 export * from './review';
