+++ apps/daengle/src/models/index.ts
@@ -1,5 +1,6 @@
 export * from './auth';
 export * from './chat';
 export * from './estimate';
+export * from './home';
 export * from './payment';
 export * from './review';
