+++ apps/daengle/src/queries/home/index.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
-import { GetUserShopsResponse, GetUserVetsResponse } from '~/models/main';
+import { getUserShops, getUserVets } from '~/apis';
+import { GetUserShopsResponse, GetUserVetsResponse } from '~/models';
 import { QUERY_KEYS } from '../query-keys';
-import { getUserShops, getUserVets } from '~/apis/main';
 
 export const useGetUserShopsQuery = () => {
   return useQuery<GetUserShopsResponse>({
