+++ packages/services/src/apis/config/index.types.ts
@@ -1,12 +1,4 @@
-import { AxiosError, AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
-
-export interface HttpClient extends AxiosInstance {
-  get<T = unknown>(url: string, data?: unknown, config?: AxiosRequestConfig): Promise<T>;
-  post<T = unknown>(url: string, data?: unknown, config?: AxiosRequestConfig): Promise<T>;
-  put<T = unknown>(url: string, data?: unknown, config?: AxiosRequestConfig): Promise<T>;
-  patch<T = unknown>(url: string, data?: unknown, config?: AxiosRequestConfig): Promise<T>;
-  delete<T = unknown>(url: string, data?: unknown, config?: AxiosRequestConfig): Promise<T>;
-}
+import { AxiosError, AxiosResponse } from 'axios';
 
 export interface DaengleError {
   status: number;
