+++ packages/services/src/constants/error-codes/index.ts
@@ -1,0 +1,4 @@
+export const ERROR_CODES = {
+  TOKEN_EXPIRED: 1003,
+  FORBIDDEN: 1101,
+} as const;
